Split airspace endpoints
